validation with reactive form adding  the worning message on the delete action

+++ src/app/customers/customers.component.ts
@@ -36,6 +36,8 @@ export class CustomersComponent implements OnInit {
   }
 
   hundelDeleteCustomer(c: Customer) {
+    let conf = confirm("Are you sure ")
+    if(!conf) return ;
     this.customerService.deleteCustomer(c.id).subscribe({
       next: (resp) => {
         this.customers = this.customers.pipe(
